# api.yaml
openapi: 3.1.0
info:
  title: GitHub Profile Enhancer API
  description: |
    GitHub 个人形象增强器 API - 用于生成和管理 GitHub 用户个人资料、头像、README 等内容的 REST API。

    ## 安全特性
    - HTTPS 强制传输
    - Bearer Token 认证
    - 请求签名支持
    - Rate Limit 限制

    ## 性能优化
    - 游标分页
    - 异步任务处理
    - 响应缓存
    - 批量操作支持

    ## 监控
    - 性能指标响应头
    - 请求追踪
    - 详细错误报告
  version: 1.0.0
  contact:
    name: API Support
    url: https://github.com/github-profile-enhancer
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.example.com/api/v1
    description: Production Server (HTTPS Required)
  - url: https://staging-api.example.com/api/v1
    description: Staging Server
security:
  - BearerAuth: []
  - RequestSignature: []
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Token 用于认证
    RequestSignature:
      type: apiKey
      in: header
      name: X-Signature
      description: |
        HMAC-SHA256 请求签名，格式: `timestamp=...;signature=...`
        签名算法: HMAC-SHA256(timestamp + request_body, api_secret)

  # 通用响应头
  headers:
...
    X-Cache-Status:
      description: 缓存状态
      schema:
        type: string
        enum: [HIT, MISS, BYPASS, EXPIRED]
        example: "HIT"
...
  parameters:
    Cursor:
      name: cursor
      in: query
      description: 游标分页标记
      schema:
        type: string
        format: base64
    PageSize:
      name: page_size
      in: query
      description: 每页数量（游标分页）
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Fields:
      name: fields
      in: query
      description: 按需返回字段（逗号分隔）
      schema:
        type: string
        example: "id,name,description,language,stars"

  # 通用 Schema
  schemas:
...
    ApiResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          description: 请求是否成功
        data:
          type: object
          description: 响应数据（成功时）
        error:
          $ref: '#/components/schemas/ErrorResponse'
        meta:
          $ref: '#/components/schemas/ResponseMeta'

    # 错误响应
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: 错误代码
          enum:
            - VALIDATION_FAILED
            - INVALID_REQUEST
            - UNAUTHORIZED
            - INVALID_TOKEN
            - TOKEN_EXPIRED
            - FORBIDDEN
            - NOT_FOUND
            - RATE_LIMIT_EXCEEDED
            - EXTERNAL_API_ERROR
            - GITHUB_API_ERROR
            - OPENAI_API_ERROR
            - INTERNAL_ERROR
            - DATABASE_ERROR
            - SIGNATURE_INVALID
            - WEBHOOK_UNVERIFIED
        message:
          type: string
          description: 错误描述信息
        details:
          type: object
          description: 错误详细信息
          properties:
            field:
              type: string
              description: 具体字段错误
            retryable:
              type: boolean
              description: 是否可重试
            retryAfter:
              type: integer
              description: 建议重试时间（秒）
            externalCode:
              type: string
              description: 外部服务错误代码
        stackTrace:
          type: string
          description: 堆栈跟踪（仅开发环境）

    # 响应元数据
    ResponseMeta:
      type: object
      required:
        - timestamp
        - requestId
      properties:
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 时间戳
        requestId:
          type: string
          format: uuid
          description: 请求 ID
        version:
          type: string
          description: API 版本

    # 分页响应（游标分页）
    CursorPagination:
      type: object
      properties:
        next_cursor:
          type: string
          format: base64
          description: 下一页游标
          nullable: true
        prev_cursor:
          type: string
          format: base64
          description: 上一页游标
          nullable: true
        has_more:
          type: boolean
          description: 是否有更多数据
        total:
          type: integer
          description: 总数（估算，非精确）

    # 异步任务响应
...
    UserInfo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        githubId:
          type: string
        username:
          type: string
        displayName:
          type: string
        email:
          type: string
          format: email
        bio:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        blog:
          type: string
          format: uri
          nullable: true
        company:
          type: string
          nullable: true
        twitterUsername:
          type: string
          nullable: true
        avatarUrl:
          type: string
          format: uri
        followers:
          type: integer
        following:
          type: integer
        publicRepos:
          type: integer
        stars:
          type: integer
        languages:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # 仓库信息
    Repository:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        fullName:
          type: string
        description:
          type: string
          nullable: true
        language:
          type: string
          nullable: true
        stars:
          type: integer
        forks:
          type: integer
        isFork:
          type: boolean
        topics:
          type: array
          items:
            type: string
        homepage:
          type: string
          format: uri
          nullable: true
        htmlUrl:
          type: string
          format: uri
        updatedAt:
          type: string
          format: date-time

    # 语言统计
...
paths:
...
  /user/info:
    get:
      tags:
        - User
      summary: 获取完整用户信息
      description: |
        获取用户完整信息，支持缓存和字段过滤。

        **缓存策略**：
        - 默认缓存时间：5 分钟
        - 使用 `?refresh=true` 强制刷新缓存
      parameters:
        - $ref: '#/components/parameters/Fields'
        - name: refresh
          in: query
          schema:
            type: boolean
            default: false
          description: 是否强制刷新缓存
        - name: include
          in: query
          description: 包含额外信息
          schema:
            type: string
            enum: [basic, repos, languages, stats, all]
            default: basic
      responses:
        '200':
          description: 成功
          headers:
            X-Cache-Status:
              $ref: '#/components/headers/X-Cache-Status'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserInfo'
  /user/repos:
    get:
      tags:
        - User
      summary: 获取用户仓库列表（游标分页）
      description: |
        使用游标分页获取用户仓库列表，支持字段过滤。

        **分页方式**：游标分页（Cursor-based Pagination）
        - `cursor`: 前一页返回的 next_cursor
        - `page_size`: 每页数量 (1-100)
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Fields'
        - name: sort
          in: query
          schema:
            type: string
            enum: [updated, stars, forks, created, name]
            default: updated
        - name: type
          in: query
          schema:
            type: string
            enum: [all, owner, member]
            default: all
        - name: direction
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          repos:
                            type: array
                            items:
                              $ref: '#/components/schemas/Repository'
                          pagination:
                            $ref: '#/components/schemas/CursorPagination'
...
  /readme/generate:
    post:
      tags:
        - README
      summary: 生成 README
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userInfo
                - config
              properties:
                userInfo:
                  $ref: '#/components/schemas/UserInfo'
                bio:
                  type: string
                avatarUrl:
                  type: string
                  format: uri
                skills:
                  type: array
                  items:
                    type: string
                projects:
                  type: array
                  maxItems: 10
                  items:
                    type: object
                config:
                  type: object
                  properties:
                    sections:
                      type: array
                    theme:
                      type: string
                      enum: [light, dark, auto]
                    showStats:
                      type: boolean
                    showVisitors:
                      type: boolean
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          readme:
                            type: string
                          wordCount:
                            type: integer
                          lineCount:
                            type: integer
  /readme/preview:
    post:
      tags:
        - README
      summary: 预览 README
      requestBody:
        $ref: '#/paths/~1readme~1generate/post/requestBody'
      responses:
        '200':
          description: 成功

  # ==================== 批量操作接口 ====================
...
  /sync/profile:
    post:
      tags:
        - Sync
      summary: 同步个人资料到 GitHub
      description: |
        将生成的个人资料同步到 GitHub 个人资料。

        **并发控制**：使用 optimistic locking 避免冲突
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bio:
                  type: string
                location:
                  type: string
                blog:
                  type: string
                  format: uri
                company:
                  type: string
                twitterUsername:
                  type: string
      responses:
        '200':
          description: 同步成功
        '409':
          description: 资料已被其他客户端修改，需要重新获取
          headers:
            X-Conflict-Token:
              description: 冲突标记
              schema:
                type: string
  /sync/readme:
    post:
      tags:
        - Sync
      summary: 同步 README 到 GitHub
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - owner
                - repo
                - content
              properties:
                owner:
                  type: string
                repo:
                  type: string
                content:
                  type: string
                commitMessage:
                  type: string
                branch:
                  type: string
                  default: "main"
      responses:
        '200':
          description: 同步成功
  /sync/pull-request:
    post:
      tags:
        - Sync
      summary: 创建 Pull Request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: PR 创建成功

  # ==================== 导出接口 ====================
...